refactor: add global root styles to index.css and move components into folders
